+++ xchange-anx/src/test/java/org/knowm/xchange/anx/v2/dto/marketdata/FullDepthJSONTest.java
@@ -3,7 +3,6 @@ package org.knowm.xchange.anx.v2.dto.marketdata;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.io.InputStream;
-import java.math.BigDecimal;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -26,15 +25,15 @@ public class FullDepthJSONTest {
 
     Assert.assertEquals(new Long(1392693096241000L), anxFullDepth.getMicroTime());
     Assert.assertEquals(3, anxFullDepth.getAsks().size());
-    Assert.assertEquals(new BigDecimal("3260.40000"), anxFullDepth.getAsks().get(0).getPrice());
+    Assert.assertEquals(new Double("3260.40000"), anxFullDepth.getAsks().get(0).getPrice());
     Assert.assertEquals(326040000L, anxFullDepth.getAsks().get(0).getPriceInt());
-    Assert.assertEquals(new BigDecimal("16.00000000"), anxFullDepth.getAsks().get(0).getAmount());
+    Assert.assertEquals(new Double("16.00000000"), anxFullDepth.getAsks().get(0).getAmount());
     Assert.assertEquals(1600000000L, anxFullDepth.getAsks().get(0).getAmountInt());
 
     Assert.assertEquals(4, anxFullDepth.getBids().size());
-    Assert.assertEquals(new BigDecimal("2000.00000"), anxFullDepth.getBids().get(0).getPrice());
+    Assert.assertEquals(new Double("2000.00000"), anxFullDepth.getBids().get(0).getPrice());
     Assert.assertEquals(200000000L, anxFullDepth.getBids().get(0).getPriceInt());
-    Assert.assertEquals(new BigDecimal("3.00000000"), anxFullDepth.getBids().get(0).getAmount());
+    Assert.assertEquals(new Double("3.00000000"), anxFullDepth.getBids().get(0).getAmount());
     Assert.assertEquals(300000000L, anxFullDepth.getBids().get(0).getAmountInt());
   }
 }
